Improve keyboard layout detection on X11
When the keyboard configuration includes multiple layouts in separate
groups, the _XKB_RULES_NAME property lists the layouts separated by a
comma. Previously we did not handle this.

We query the current keyboard group on startup and monitor
XkbStateNotify events for group changes. The existing infrastructre
from QKeyMapper deals with sending the KeyboardLayoutChange event to
all toplevel windows already.

This change includes a bit of refactoring to put the XKB detection and
initialization in qapplication_x11.cpp and XKB opcode, eventbase, and
errorbase into qt_x11_p.h (like with the other extensions).

Task-number: QTBUG-3631
Reviewed-by: denis