Improve performance of QScriptValueIterator
The implementation was getting all the property names
(identifiers) and storing their string representation.
In e.g. value() and setValue(), the string representation
then had to be converted back to an identifier. This was
unnecessary work.

Instead, store the identifiers directly, and use them
via the QScriptValue private API whenever we can.
This greatly improves the performance of scriptName(), value(),
setValue(), flags() and remove().
(tests/benchmarks/script/qscriptvalueiterator is 4-5 times faster.)

Reviewed-by: Jedrzej Nowacki