Cocoa: Fix misaligned text between format changes
This reverts part of commit e008504b5ec34975e34adf3b1a2b7170d0e4dd38.
The original commit aimed to round up all fractional numbers for the
font to "improve appearance". Rounding up the last advance of a set
of advances in a glyph layout, however, breaks text layout completely,
since the layout of text will be different depending on how many text
lines is used to represent it (since the last glyph in each line will
have a higher advance than the others.) This was e.g. visible in
Qt Creator when turning on whitespace visualization, as all spaces
then created a format change, thus a new text line in the text layout,
and hence an increase of X pixels in the positioning of the text after
the space, where 0.0 < X < 1.0. I see no negative results of removing
the rounding, but if there are any problems, it has to be fixed in the
usage of the advances not in the advances themselves.

Task-number: QTCREATORBUG-381, QTBUG-8791
Reviewed-by: mae