Fixed dialog resize not to move the dialog for Symbian.
QDialog::resize() also moved the dialog in Symbian. This occured since
adjustPosition was called as an result of resize and that method did
not check if dialog position was explicitly set.

In addition it was found that in Symbian WA_Resized and WA_Moved
attributes were basically set for almost all top-level widgets
by system. This was also fixed and a new auto test was introduced
to verify these attributes in all platforms. Windows platform also
suffers from bug in this area, and a separate task QTBUG-5897 for
it was created

Task-number: QTBUG-5897
Reviewed-by: Sami Merila