qmake: fix to create proper install target in Makefile at first run
Previously when qmake was executed the first time, or when no
previously compiled binary target was available for some reason,
qmake just created an empty install target in the Makefile.
So in fact for a target to be installed properly you needed to run
something like "qmake && make && qmake && make install".
This should now be fixed with this patch.

Reviewed-by: Marius Storm-Olsen
Task-number:QTBUG-5558