Slow QTreeView::layout() with many (> 10000) children
When updating the QTreeViewItem::total field in layout(), we used to
call QTreeViewPrivate::viewIndex() to get the parent item, which is
O(n). We have now introduced 'parentItem' in QTreeViewItem wich makes
this O(1), with a small penality when inserting and removing elements
in QTreeViewPrivate::viewItems.

The QTreeViewPrivate::checkViewItems() is left inside the code for
further debugging.

Reviewed-by: Olivier
Task-number: QTBUG-8885
(cherry picked from commit 7baaea978200c82fdf65e3934cfa373edeca6524)