Improve value type binding behavior
Changing value type bindings in state changes, and implicitly removing
them on property assignment was not reliable.  Internally the system
considered a binding on "font" and one on "font.x" as a binding on two
separate properties, even though the "font" binding completely
overrides the "font.x" property.

Following this change a binding to "font.x" creates a proxy binding object
on the "font" property in addition to the "font.x" binding itself.  This
allows behavior to be consistent across all operations.

QT-2920