Mac: scroll speed does not match native applications
It turns out that we limit wheel scrolling on mac to an
upper limit per event to one page. This behaviour is a bit
strange in the first place, and on mac, it is just wrong.
Besides, even when this limitation is removed, we still scroll
a bit slower that native. The 'problem' is to come up with a
good conversion from pixel scrolling to line based
scrolling (Qt does not really have an API for pixel scrolling).
But we up the speed a bit to make it perform more similar to
xcode.

Rev-by: msorvig