Fixed problem displaying half width character as full width
There is the problem that Qt displays half width character
(e.g. alphabet) as full width. This is reported in #QTBUG-1726.

This occurs in the following fonts.
 - Monospace font
 - Selif font
 - IPA font (Ver.0301)
 - NSimSun

In Cario and fontconfig, the FT_LOAD_IGNORE_GLOBAL_ADVANCE_WIDTH
is specified as flag of FT_Load_Glyph function to solve this problem.
According to the document of freetype, this flag exists to support
buggy CJK fonts.

So this flag is specified as default flag of QFontEngineFT.

Merge-request: 2319
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@nokia.com>