Minisplitter doesn't paint to the bottom/right
The problem is not related to the Splitter handle but related to an
optimization in the paintengine.
We were comparing if the size of the widget had changed before setting
the new mask, which might not be correct. We could have changed the mask
without changing the widget size.

Task-number: QTCREATORBUG-753
Reviewed-by: Morten
Reviewed-by: Samuel