HotFix for fluidlauncher default size caused by 6d44dadd.
It seems that in Symbian we cannot restore Qt::WA_Moved and
Qt::WA_Resized attributes in setWindowState method to the same ones
what the attributes were when method was called. Thus we currently
make sure that Qt::WA_Moved and Qt::WA_Resized attributes are not
touched when normal window state is applied.

There is a new task QTBUG-8977 to sort out when those attributes
should be set and when not.

Reviewed-By: Sami Merila