Moved softkey "Options" action handling from QMainWindow to QMenuBar.
This change is done to enable QMenuBar usage and related "Options"
softkey also on other than QMainWindow objects. For example with this
change it is possible to create menubar and set it to QLayout by using
setMenuBar, and it will then be usable via "Options" softkey.

This functionality makes it easy to use different menus for different
views in QStackedWidget.

Task-number: QT-2275        (Not yet finished, but related)
Reviewed-by: Sami Merila
Reviewed-by: Alessandro Portale