Make QColorDialog usable on small screens
QColorDialog was always assuming that it got its minimal size requirements.
On a lot of screens, this was not the case, so choosing a color in the
QColorPicker was totally off. This patch handles the resize event correctly,
so the QColorPicker will be usable on all possible screen sizes.

This fixes Maemo's bug https://bugs.maemo.org/show_bug.cgi?id=9526

Reviewed-by: Robert Griebl