Implement a special case, simplified vertex shader for complex geometry
When drawing text, the vertex count will most likely be so high that
using a uniform-based, simpler vertex shader is faster. We implement
the ability to inform the shader manager that the geometry is considered
complex, so that it can choose the simpler vertex shader in these cases.

Task-number: QT-2887
Reviewed-by: tom