QXmlSchema internals: fix crash with anonymous types
The crash occurred when an anonymous type was created that was a list
type. In that case, we did not set the item type, which is used later at
XsdSchemaChecker::checkSimpleDerivationRestrictions(), which would lead
to a crash.

Additionally, in the xmlpatternsvalidator test, check the exit status of
the process after it has finished, to detect crashes.

Reviewed-by: Tobias Koenig <tokoe@kde.org>
Reviewed-by: Frans Englich
Task-number: QTBUG-8920