Improved qt_x11_wait_for_window_manager
We shouldn't really care about ReparentNotify when waiting for a window to be
"managed" by the window manager.

We only need to wait until the window is Mapped and Exposed for the first time,
then we (most likely) can safely assume that the window manager has finished
managing the window.

Task-number: QTBUG-9097
Reviewed-by: Bradley T. Hughes