Fixes for the module import
The change makes import "url" also load plugins. We know must
decide on the preferred coding style.

Since native types must be registered with a unique uri, we now
generate a stable uri from the plugin path and the import paths.
This way we do not have to call registerTypes() multiple times.