"Strictly" Fullscreen Application in Mac OS
Full screen in OSX is achieved by using the Kiosk Mode. Therefore it is
not a real full screen but an emulation since we basically disable the
dock and the menubar. Previous to this patch we were autoshowing the
menubar all the time even when we didn't have a menubar. The problem
with that is the fact that in OSX there is always a system menubar,
so the menubar appeared the moment somebody hovered over the menubar
region. With this patch we check if we have created a menubar and if so
we autoshow the menubar, otherwise the menubar will remain hidden in
full screen mode.

Task-number: QTBUG-8933
Reviewed-by: Richard Moe Gustavsen