Useful support for -qtlibinfix configure parameter in Symbian
Configure parameter -qtlibinfix will now change also plugin names
of plugins built and installed with Qt in Symbian.
Default plugin path is changed to: /resource/qt<libinfix>/plugins.
Other plugins besides the ones installed with Qt are not renamed.

With infixed configuration, Qt can be installed on a phone that already
has a Qt installed on ROM without interfering with the ROM version of
Qt.

Note that since s60main.rsc resource cannot deployed with infixed Qt,
and infixing it is somewhat problematic, currently the phone needs
to have proper Qt installation, too, for infixed Qt to work.
It also means that any changes to Qt that would affect s60main.rsc
cannot be tested on real device using infixed builds. Since this file
is unlikely to need changing, this should not be a big problem.

Task-number: QTBUG-9065
Reviewed-by: Jani Hautakangas