Fixes blending problem when paiting non-opaque items with cache enabled.
In most cases when partially updating an item, the old content of the
cache needs to be replaced with the new content
(CompositionMode_Source). But in a specific case when using
DeviceCoordinateCache and when the item is transformed, the new content
needs to be blended (CompositionMode_SourceAtop) with the old one to
avoid incorrect fully transparent background.

Autotest included.

Task-number: QTBUG-7863
Reviewed-by: bnilsen