Optimize getting bearings of a glyph on Windows for true type fonts
Usually, the bearing of a glyph is calculated based on the bounding box
of the glyph. However, retrieving these bounds is a costly operation
on Windows. For true type fonts, we can retrieve the bearings directly
by getting the ABC widths, so as an optimization we implement a new
virtual function in QFontEngine which retrieves the bearings of the
glyph. The default implementation uses boundingBox, but on Windows we
try to take the shortcut.

Reviewed-by: Simon Hausmann