Add a raw bind() function to QGLBuffer.
It is awkward for a Qt application to do the equivalent of
glBindBuffer(GL_ARRAY_BUFFER, 0) without knowing the QGLBuffer
that was previously bound, or to bind a raw id obtained elsewhere.
Resolving the extension is annoying.

This change provides a raw low-level version of bind() for directly
resolving and calling glBindBuffer() to assist such applications.

Reviewed-by: Sarah Smith