Optimized visualRegionForSelection in various item views classes
When the number of selected items is large (a few thousands), the
computation of visualRegionForSelection can take a long time (up to a
few seconds). Analysis with valgrind shows that most of the time is
spent in miRegionOp (in qregion.cpp), which is being called by
QRegion::operator+=. The visualRegionForSelection virtual method being
called only to update the view after selection, we can safely ignore
those item's rectangles outside the viewport, thus both reducing the
number of calls to miRegionOp and the actual cost of each call.

This, however, introduces a behaviour change in
visualRegionForSelection, as the returned region will *not* contain
any rectangle *not* intersecting the viewport.

Reviewed-by: Thierry
Task-number: QTBUG-884