Redesigned filter widgets
I redesigned the filter widget so that the reset button is contained
within the line edit and only shown when edited. This is typically
what KDE and GNOME does. The base icon also changed to use platform
dependant icons on X11 and a new clear icon for other platforms.

Since hinttext is now a feature of QLineEdit, I could safely
remove this code as well.

Unfortunately Oxygen is reporting incorrect contents margins for
line edits. In essence it incorrectly offsets the frame so I had
to add some magic to support that in particular.

Reviewed-by: ogoffart