Initialize the graphics system before creating the style.
Previously if the style was creating any pixmaps they were going to the
createDefaultPixmapData() function which creates a default platform
pixmap without considering the graphics system. If these pixmaps are
used they may end up going through a suboptimal code path because they
are not the same type of pixmap that is expected by the graphics
system. The fix here is to intialize the graphics system before the
style is created.

Reviewed-by: Trond