Add a selftest to enforce correct test naming conventions.
When a test runs successfully, it outputs an XML test log, with the
testcase name coming from whatever class is passed to QTest::qExec.

If a test crashes, hangs, fails to compile, or otherwise fails to output
a valid test log, any summary of the test results still has to display
some reasonable testcase name.  But we can't tell what the testcase name
_would_ have been had the test run correctly - without, for example,
attempting to parse the C++.

The simplest solution is to make sure that TARGET and the test class
name are always matching.  This test will force people to follow this
naming convention.