Redesigned object and class labels for the Designer propertyeditor
We modified the label so that it gets its own row instead of
squeezing it into the tool bar. This layout would usually not
have room for more than short labels and looked rather
unprofessional when used in Creator. The new label is properly
elided. We also added a frame to distinguish it from the dock
widget label.

Reviewed-by: Friedemann Kleint