Adding custom pixel metrics requires cleaning and rebuilding of QtGui
This is due that the size of the pixel metrics table is stored in some
object file and the file is not updated, even if we increase the table
size integer. Removing direct calls from internal methods to member
variable data[] removes the need to do whole rebuild when adding
values.

Relates to task #9247.

Task-number: QTBUG-9247
Reviewed-by: Alessandro Portale