Cocoa: fix eventdispatcher crash, found by macgui autotest
In some cases, we end up deleting a widget before we get to
end the modal session that might be attached to it. And that
will cause a crash. This patch retains/releases the underlying
NSWindow, so we always have a valid reference as long as the
modal session is alive.

Reviewed-by: msorvig