Fix #ifdef logic for Symbian conversion functions in QVGPixmapData.
The previous #ifdef logic had the entire body of fromNativeType() and
toNativeType() #ifdef'ed out meaning that the OpenVG graphics system
could not convert CFbsBitmap instances to VGImage instances. This was
obviously incorrect because this code has no dependancy on RSgImage (or
EGL) and should therefore be outside of the #ifdef.

Reviewed-by: TrustMe