Fixing keypad navigation focus frame
The QFocusFrame drawing code in QS60Style incorrectly assumed that
PM_LayoutSpacing is always equal to PM_FocusFrameMargin. When these
values diverged, the focus frame broke as described in task QTBUG-8036.

The fix makes the drawing more robust. The focus frame width is never
thicker than PM_LayoutSpacing and PM_FocusFrameMargin. And instead of
drawing a roundRect with calculated pen width, we are now filling a
roundRect. That makes the focusFrame more apparent.

Task-number: QTBUG-8036
Reviewed-by: Sami Merila