Fixed S60 softkey implementation to use popup/modal dialog softkeys.
Previously softkey implementation assumed that source for softkey
actions is focused widget or activewindow if there is no focused
widget. Since for example pop-up menu does not take focus immediately
in touch enabled devices, the underlying widget softkeys were visible
and usable when menu was open. This lead to problem that underlying
widget could be interacted via softkeys when popup menu was open.

For example as reported in QTBUG-8688, it was possible to close the
underlying filedialog from which the currently active context menu was
launched. It was also possible to open "Options" menu of underlying
widget when context menu was open, leading to situation where several
context menus could be launched via "Actions" item of "Options" menu.
This was reported as an issue: QTBUG-6167

In addition when user started navigating in context menu via keypad,
the menu got focus and softkeys changed to partially correct. Only
partially correct, since context menu "Select" action and "Options"
action had same priority and both associated to LSK. In addition the
context menu action "Cancel" associated to RSK was set to invisible,
meaning that it was also disabled and appeared in sofkeys as dimmed.

All of these issues were fixed by making making the popup and modal
dialogs highest priority softkey source. In case the focused widget
is inside popup or modal dialog it is being used as an initial softkey
source. In addition the softkeys created with
QSoftKeyManager::createAction or QSoftKeyManager::createKeyedAction are
now by default invisible i.e. not visible in context menu and have
special property set to make them still normally enabled in softkeys.

Task-number: QTBUG-6167
Task-number: QTBUG-8688
Task-number: QTBUG-9144

Reviewed-by: Sami Merila