Optimize QMetaObject::indexOf* functions
Comparing the first character before calling strcmp gives big
speedup, as it avoid the overhead of calling the strcmp function
in many cases.

In indexOfMethodRelative, fix the normalize case. The computation
of the normalized string was put outside of the for loop, but it should
have been inside.  This was not detected by the test because the wrong
string is at the end.

In IndexOfMethodRelative, we do not need to check if the method is a signal
or a slot, because we only iterate over the right interval.
This is only true for code generated by moc since Qt 4.6.
Which means that in application compiled with Qt 4.5 and older,
indexOfSignal could now return a slot with the same name (and vice-versa),
but I this it is safe to ignore that "problem".

Reviewed-by: Roberto Raggi
Reviewed-by: Kent Hansen