Add new function QTextLine::horizontalAdvance()
Add a function to QTextLine which returns the accumulated advance of
the glyphs in the text. Previously QTextLine::naturalTextWidth() was
used for both this purpose and for calculating the pixel width of the
text. Since these two metrics are not the same, either of the two
usages would be wrong. QTextLine::naturalTextWidth() has been changed
to do what its documentation claims it does, and horizontalAdvance()
should now be used for laying out text horizontally.

Reviewed-by: Simon Hausmann