Mac: new auto test: macnativeevents
This test will utilize the new API I wrote back in the
days for posting/listening for native mouse and keyboard
events on mac. By being able to create low-level native events
that you can post to the window server, we can write tests that
not only test the widgets layer of Qt, but also the platform
dependent code that translates native events into Qt events.

Reviewed-by: msorvig