Do not move keyboard focus to invisible windows on X11.
When QWidget::activateWindow() is called we check if the window was already
mapped and just do not do anything if it wasn't.  We used to try to active it
by giving it keyboard focus, but that always fails because it isn't allowed to
call XSetInputFocus on invisible windows, and the error message was swalled by
Qt in the error handler. So now we just don't bother even trying if we know the
call will fail.

Task-number: QTBUG-4042
Reviewed-by: Benjamin Poulain
Reviewed-by: Bradley T. Hughes