Give 1-pixel splitters a larger grab area
This is a trick we have applied to Qt Creator for a
while and it makes sense to backport this into Qt.

Essentially 1-pixel splitters are unusable at the
moment. With this fix the actual splitter handle
will be 5 pixels wide and slightly overlap neighbouring
widgets. The layout code in QSplitter alreadu use
the contentsRect to support this behavior.

Task-number: QTBUG-9335
Reviewed-by: thierry