Tweak import behaviour
The implicit import of types in the same directory/module now imports
version -1 (which means all versions). This fixes the Twitter demo.

Additionally, the current directory is not added to the imports path.
No functionality is lost, as you can import local modules with relative
paths for the same effect.

Improved docs are on the way.

Reviewed-by: mae