Fix QFontMetrics::height() and QFontMetrics autotest
Since the Mac font engine has been fixed to support fractional metrics,
we can no longer expect qRound(A + B) to be equal to qRound(A) +
qRound(B). The documentation states that height = ascent + descent + 1.
To honor this contract, we need to rewrite the implementation of
height() to round each metric separately. Same with lineSpacing.

The bearingIncludedInBoundingrect() test was wrong. We can guarantee
that the italic text is not more narrow than the normal text, but we
cannot guarantee that they are never the same width.

Reviewed-by: Olivier