Fix QLabel::sizeHint()
The QTextControl case of QLabel calculated the bounding rect of the
label by taking the size of the control (floating point) and using
toSize() to convert this to an integer QSize. While this would work
when the font engine always returned integer values, it is not the
correct way to calculate a bounding rect, since the bounding rect needs
to contain all the pixels of the text, including the ones that are only
partially covered, thus its size should never be rounded down.

Reviewed-by: Olivier