QMessageBox is smaller than native MessageBox
Messageboxes are smaller than native ones.

This is due to that native ones have a lot of empty space.

To fix this, we define a new custom pixel metrics that is the minimum
height of one text line messagebox (aka AknPopUp) on the native side.
Then we ensure that the QMessageBox is at least of this height.

Additionally we do some minor styling for QMessageBox:
- the corners graphics are now as rounded as on native side
- the font is set to match the native side
- the top margin space is doubled for dialogs, which is rather good
approximation of native side

Task-number: QTBUG-4875
Reviewed-by: Janne Anttila