Symbian event loop priority drop not used in Symbian^4
The priority drop hack used in the Symbian version of Qt's event
dispatcher is not needed from Symbian^4, as the viewsrv panics, which
it was designed to prevent, are disabled. This change disables the
priority drop code when Qt is compiled in a Symbian^4 environment,
using the SYMBIAN_GRAPHICS_WSERV_QT_EFFECTS define as a key.

This will improve system and performance predictability, as app
priorities will be more stable. It also gives a slight performance
boost to app startup in the order of 10ms.

Reviewed-by: Jason Barron