Fix glyph advance on Mac/Carbon
Device metrics was turned on for Mac/Carbon to work around a bug
in WebKit that required the font engine to report only integer values.
This in turn caused positioning of glyphs to be slightly off compared
to native applications. The bug has been fixed in WebKit, so now we
can turn off the device metrics flag. Since we are now reporting
fractional values as well, we need to round the numbers when the
ForceIntegerMetrics style strategy flag is set.

Task-number: qtbug-5529
Reviewed-by: Trond