QtScript: Add API for reporting additional memory costs
QScriptEngine::reportAdditionalMemoryCost(int).

This function provides the ability to give a hint to the engine
that it should perhaps trigger garbage collection sooner rather
than later.

For example, if you've implemented a JS ByteArray class that
wraps a QByteArray, and a user constructs a few hundred
temporary ByteArray objects of large sizes, failure to report
the additional memory cost may cause the application's memory
consumption to grow and grow (because the script engine thinks
they are "cheap" objects, the GC won't kick in).

Reporting the correct size can be difficult (or impossible) in
some cases. For example, it's difficult to predict the total
amount of system memory & resources consumed by a QImage.
But even reporting a heuristic / approximate cost can be better
than reporting no cost.

Task-number: QTBUG-6238
Reviewed-by: Simon Hausmann