QtScript: Add yet more missing API shims
- for QScriptEngine and QScriptContext; we don't control what the
JSC functions are doing, so it's better to have shims than not to
have them (which can cause something to blow up), even if they
might not be strictly necessary as of this writing.

- for QScriptDeclarativeClass; otherwise the identifiers might
be created/destroyed in the wrong engine.

- for QScriptValueIterator destructor; otherwise the identifiers
might be destroyed in the wrong engine.

This is an attempt to fix a crash in Bauhaus, but it might still
be crashing (i.e. still some shims missing, somewhere) ;(