QS60Style: very tall QSpinBox's buttons hide lineEdit
If the spinbox is very tall considering its width (ratio 2:1), the
spinbuttons can easily hide the lineEdit when they are scaled up.

For example QSize(200, 100) QSpinBox does not have any lineEdit at all,
since spinbuttons are scaled up to be 100*100 squares and deployed
side-by-side.

As a fix, once spinbuttons occupy half of the total width of spinbox,
they stop growing. For very tall spinboxes, spinbuttons are deployed
one top of the other, to make the lineEdit bigger.

Task-number: QTBUG-9321
Reviewed-by: Alessandro Portale