Fixed bitfield-related crash on Symbian WINSCW
When running on WINSCW, the nativePaintMode variable was becoming
corrupted.  The point at which this corruption occurs is unclear,
but it can be reproduced as follows:

1. Launch qmediaplayer
2. In the Phonon::MMF::DsaVideoOutput constructor, the
   nativePaintMode flag is set to QWExtra::ZeroFill
3. Open a video clip
4. During start of playback, QSymbianControl::Draw is called on the
   native control corresponding to the DsaVideoOutput object.  This
   checks the value of nativePaintMode; it does not match any of
   the valid NativePaintMode values, so an assertion fails.

This crash does not occur on target, suggesting that the cause may
be an error in the WINSCW compiler.  Although the C++ standard allows
bitfields to have boolean, integral or enumeration type, the latter
is not permitted by C99.

Neither increasing the number of bits allocated to nativePaintMode,
nor changing it position in the list of platform-specific bitfields
has any effect.

After making nativePaintMode into a normal field rather than a
bitfield, the crash no longer happens.  Note that, since bitfields
must be kept together, nativePaintMode is moved to the end of the
structure.