Fix QFileSystemModel to not install useless watchers on the filesystem
When someone call setRootPath we need to remove the previous watcher.
This save memory and also it avoids Qt Creator to run out of filesystem
watcher which then leads to a crash. If you call setRootPath on the
previous path all changes that might happen in between will be propagated
in the model.

Task-number:QTBUG-8632
Reviewed-by:Friedemann Kleint